+++ CraftyServer/Core/ItemBoat.cs
@@ -12,10 +12,10 @@ namespace CraftyServer.Core
             float f = 1.0F;
             float f1 = entityplayer.prevRotationPitch + (entityplayer.rotationPitch - entityplayer.prevRotationPitch)*f;
             float f2 = entityplayer.prevRotationYaw + (entityplayer.rotationYaw - entityplayer.prevRotationYaw)*f;
-            double d = entityplayer.prevPosX + (entityplayer.posX - entityplayer.prevPosX)*(double) f;
-            double d1 = (entityplayer.prevPosY + (entityplayer.posY - entityplayer.prevPosY)*(double) f +
-                         1.6200000000000001D) - (double) entityplayer.yOffset;
-            double d2 = entityplayer.prevPosZ + (entityplayer.posZ - entityplayer.prevPosZ)*(double) f;
+            double d = entityplayer.prevPosX + (entityplayer.posX - entityplayer.prevPosX)*f;
+            double d1 = (entityplayer.prevPosY + (entityplayer.posY - entityplayer.prevPosY)*f +
+                         1.6200000000000001D) - entityplayer.yOffset;
+            double d2 = entityplayer.prevPosZ + (entityplayer.posZ - entityplayer.prevPosZ)*f;
             Vec3D vec3d = Vec3D.createVector(d, d1, d2);
             float f3 = MathHelper.cos(-f2*0.01745329F - 3.141593F);
             float f4 = MathHelper.sin(-f2*0.01745329F - 3.141593F);
@@ -25,7 +25,7 @@ namespace CraftyServer.Core
             float f8 = f6;
             float f9 = f3*f5;
             double d3 = 5D;
-            Vec3D vec3d1 = vec3d.addVector((double) f7*d3, (double) f8*d3, (double) f9*d3);
+            Vec3D vec3d1 = vec3d.addVector(f7*d3, f8*d3, f9*d3);
             MovingObjectPosition movingobjectposition = world.rayTraceBlocks_do(vec3d, vec3d1, true);
             if (movingobjectposition == null)
             {
@@ -38,7 +38,7 @@ namespace CraftyServer.Core
                 int k = movingobjectposition.blockZ;
                 if (!world.singleplayerWorld)
                 {
-                    world.entityJoinedWorld(new EntityBoat(world, (float) i + 0.5F, (float) j + 1.5F, (float) k + 0.5F));
+                    world.entityJoinedWorld(new EntityBoat(world, i + 0.5F, j + 1.5F, k + 0.5F));
                 }
                 itemstack.stackSize--;
             }
